+++ xray/core/dataarray.py
@@ -120,7 +120,7 @@ class DataArray(AbstractArray, BaseDataObject):
     attrs : OrderedDict
         Dictionary for holding arbitrary metadata.
     """
-    groupby_cls = groupby.ArrayGroupBy
+    groupby_cls = groupby.DataArrayGroupBy
 
     def __init__(self, data, coords=None, dims=None, name=None,
                  attrs=None, encoding=None):
@@ -700,6 +700,30 @@ class DataArray(AbstractArray, BaseDataObject):
         ds = self._dataset.dropna(dim, how=how, thresh=thresh)
         return self._with_replaced_dataset(ds)
 
+    def fillna(self, value):
+        """Fill missing values in this object.
+
+        This operation follows the normal broadcasting and alignment rules that
+        xray uses for binary arithmetic, except the result is aligned to this
+        object (``join='left'``) instead of aligned to the intersection of
+        index coordinates (``join='inner'``).
+
+        Parameters
+        ----------
+        value : scalar, ndarray or DataArray
+            Used to fill all matching missing values in this array. If the
+            argument is a DataArray, it is first aligned with (reindexed to)
+            this array.
+
+        Returns
+        -------
+        DataArray
+        """
+        if utils.is_dict_like(value):
+            raise TypeError('cannot provide fill value as a dictionary with '
+                            'fillna on a DataArray')
+        return self._fillna(value)
+
     def reduce(self, func, dim=None, axis=None, keep_attrs=False, **kwargs):
         """Reduce this array by applying `func` along some dimension(s).
 
@@ -930,13 +954,13 @@ class DataArray(AbstractArray, BaseDataObject):
         return func
 
     @staticmethod
-    def _binary_op(f, reflexive=False):
+    def _binary_op(f, reflexive=False, join='inner', **ignored_kwargs):
         @functools.wraps(f)
         def func(self, other):
             if isinstance(other, (Dataset, groupby.GroupBy)):
                 return NotImplemented
             if hasattr(other, 'indexes'):
-                self, other = align(self, other, join='inner', copy=False)
+                self, other = align(self, other, join=join, copy=False)
                 empty_indexes = [d for d, s in zip(self.dims, self.shape)
                                  if s == 0]
                 if empty_indexes: