+++ xray/core/variable.py
@@ -578,6 +578,9 @@ class Variable(common.AbstractArray):
                                 self._encoding, fastpath=True)
         return expanded_var.transpose(*dims)
 
+    def fillna(self, value):
+        return self._fillna(value)
+
     def reduce(self, func, dim=None, axis=None, keep_attrs=False,
                **kwargs):
         """Reduce this array by applying `func` along some dimension(s).
@@ -784,7 +787,7 @@ class Variable(common.AbstractArray):
         return func
 
     @staticmethod
-    def _binary_op(f, reflexive=False):
+    def _binary_op(f, reflexive=False, **ignored_kwargs):
         @functools.wraps(f)
         def func(self, other):
             if isinstance(other, (xray.DataArray, xray.Dataset)):