+++ xray/test/test_dataarray.py
@@ -803,6 +803,43 @@ class TestDataArray(TestCase):
         self.assertEqual(len(vm.attrs), len(self.attrs))
         self.assertEqual(vm.attrs, self.attrs)
 
+    def test_fillna(self):
+        a = DataArray([np.nan, 1, np.nan, 3], dims='x')
+        actual = a.fillna(-1)
+        expected = DataArray([-1, 1, -1, 3], dims='x')
+        self.assertDataArrayIdentical(expected, actual)
+
+        b = DataArray(range(4), dims='x')
+        actual = a.fillna(b)
+        expected = b.copy()
+        self.assertDataArrayIdentical(expected, actual)
+
+        actual = a.fillna(range(4))
+        self.assertDataArrayIdentical(expected, actual)
+
+        actual = a.fillna(b[:3])
+        self.assertDataArrayIdentical(expected, actual)
+
+        actual = a.fillna(b[:0])
+        self.assertDataArrayIdentical(a, actual)
+
+        with self.assertRaisesRegexp(TypeError, 'fillna on a DataArray'):
+            a.fillna({0: 0})
+
+        with self.assertRaisesRegexp(ValueError, 'broadcast'):
+            a.fillna([1, 2])
+
+        fill_value = DataArray([0, 1], dims='y')
+        actual = a.fillna(fill_value)
+        expected = DataArray([[0, 1], [1, 1], [0, 1], [3, 3]], dims=('x', 'y'))
+        self.assertDataArrayIdentical(expected, actual)
+
+        expected = b.copy()
+        for target in [a, expected]:
+            target.coords['b'] = ('x', [0, 0, 1, 1])
+        actual = a.groupby('b').fillna(DataArray([0, 2], dims='b'))
+        self.assertDataArrayIdentical(expected, actual)
+
     def test_groupby_iter(self):
         for ((act_x, act_dv), (exp_x, exp_ds)) in \
                 zip(self.dv.groupby('y'), self.ds.groupby('y')):
@@ -952,6 +989,17 @@ class TestDataArray(TestCase):
             grouped + 1
         with self.assertRaisesRegexp(TypeError, 'only support arithmetic'):
             grouped + grouped
+
+    @unittest.skip
+    def test_groupby_math_not_aligned(self):
+        # We need to fix Variable.concat to infer dtypes properly before this
+        # will pass. For now, raising KeyError when labels are missing (instead
+        # of aligning) is a reasonable fallback.
+        array = DataArray(range(4), {'b': ('x', [0, 0, 1, 1])}, dims='x')
+        other = DataArray([10], dims='b')
+        actual = array.groupby('b') + other
+        expected = DataArray([10, 11, np.nan, np.nan], array.coords)
+        self.assertDataArrayIdentical(expected, actual)
 
     def test_groupby_restore_dim_order(self):
         array = DataArray(np.random.randn(5, 3),